Rename db connection related file name & rename domain config parameters
