+++ src/index.ts
@@ -1,6 +1,11 @@
 // Core config & connections
-export { initMultiTenantConfig, getConfig } from "./core/config";
-export { connectToTenantsDb, getTenantsConnection, getTenantDbConnection, getTenantUri } from "./core/core";
+export { initMultiTenantConfig, getConfig } from "./tanent-db/config";
+export {
+  connectToTenantsDb,
+  getTenantsConnection,
+  getTenantDbConnection,
+  getTenantUri,
+} from "./tanent-db/dbConnection";
 
 // for apps that want to resolve tenant based on host
 export { resolveTenant } from "./utils/resolveTenant";
